refactor pubsub message handling and add PushKind type with string conversion
