+++ go/integTest/PubSubTestPlan.md
@@ -7,6 +7,7 @@ This document outlines the integration test plan for PubSub functionality in Val
 **Current Status**: PubSub commands are not fully implemented in the Valkey GLIDE Go wrapper, so we'll use the CustomCommand interface for testing.
 
 **Implementation Strategy**:
+
 1. Complete all tests for GlideClient first
 2. Then implement the same tests for GlideClusterClient
 3. Focus on one feature at a time, iterating until it works correctly before moving to the next
@@ -20,23 +21,29 @@ Tests will be organized in a separate test suite (PubSubTestSuite) to allow comp
 ### GlideClient PubSub Tests
 
 #### Basic Publish/Subscribe
+
 - [ ] Test basic publish and subscribe functionality
+  - In Progress
+  - Working to resolve issues with message delivery for multiple subscribers.
 - [ ] Test message delivery to multiple subscribers
 - [ ] Test unsubscribe functionality
 - [ ] Test message pattern matching with PSUBSCRIBE
 
 #### Message Handler
+
 - [ ] Test message handler receives published messages
 - [ ] Test message handler with multiple channels
 - [ ] Test message handler with pattern subscriptions
 - [ ] Test error handling in message handler
 
 #### Subscription Configuration
+
 - [ ] Test subscription with different buffer sizes
 - [ ] Test subscription timeout configuration
 - [ ] Test reconnection behavior with subscriptions
 
 #### Edge Cases
+
 - [ ] Test behavior with high message volume
 - [ ] Test behavior when Redis server disconnects
 - [ ] Test message ordering guarantees
@@ -44,23 +51,27 @@ Tests will be organized in a separate test suite (PubSubTestSuite) to allow comp
 ### GlideClusterClient PubSub Tests
 
 #### Basic Publish/Subscribe
+
 - [ ] Test basic publish and subscribe functionality
 - [ ] Test message delivery to multiple subscribers
 - [ ] Test unsubscribe functionality
 - [ ] Test message pattern matching with PSUBSCRIBE
 
 #### Message Handler
+
 - [ ] Test message handler receives published messages
 - [ ] Test message handler with multiple channels
 - [ ] Test message handler with pattern subscriptions
 - [ ] Test error handling in message handler
 
 #### Subscription Configuration
+
 - [ ] Test subscription with different buffer sizes
 - [ ] Test subscription timeout configuration
 - [ ] Test reconnection behavior with subscriptions
 
 #### Edge Cases
+
 - [ ] Test behavior with high message volume
 - [ ] Test behavior when cluster nodes disconnect
 - [ ] Test message ordering guarantees
@@ -75,12 +86,6 @@ Current focus: GlideClient Basic Publish/Subscribe tests
 - [ ] UNSUBSCRIBE command is not fully implemented in the Rust core (lib.rs:574 "not yet implemented")
   - The `TestUnsubscribeWithGlideClient` test is currently skipped due to this limitation
   - Need to revisit once the Rust implementation is updated
-- [ ] Multiple subscribers test required significant refactoring:
-  - Using separate channels for each subscriber to avoid potential conflicts
-  - Using separate subscription channels for each client to ensure message delivery
-  - Added mutex-protected flags to track message receipt
-  - Continuous publishing of messages until both subscribers receive at least one
-  - Added more detailed logging to help diagnose issues
 
 ## Future Refactoring Tasks
